requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dateutil==2.9.0.post0
jinja2==3.1.6
psycopg[binary]==3.2.9
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
